+++ shift/system/core/http/router.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shift\System\Core\Http;
+
+class Router
+{
+    protected $urlRewrite = [];
+
+    public function addUrlRewrite(object $urlRewrite)
+    {
+        if (method_exists($urlRewrite, 'urlAlias')) {
+            $this->urlRewrite[] = $urlRewrite;
+        }
+    }
+
+    public function url(string $route, string $args = '', int $language_id = 0): string
+    {
+        foreach ($this->urlRewrite as $rewriter) {
+            if ($url = $rewriter->urlAlias($route, $args, $language_id)) {
+                return $url;
+            }
+        }
+
+        return URL_APP . 'index.php?route=' . $route . ($args ? '&' . trim($args, '&') : '');
+    }
+}